+++ ProdManager/routes/incident/forms.py
@@ -31,7 +31,8 @@ class IncidentCreateForm(CustomForm):
   description = TextAreaField(
     name='description',
     label=lang.get("table_column_description").capitalize(),
-    validators=[Optional()]
+    validators=[Optional()],
+    render_kw=dict(placeholder=lang.get("placeholder_input_description").capitalize())
   )
   external_reference= StringField(
     name='external_reference',
@@ -89,7 +90,8 @@ class IncidentCommentForm(CustomForm):
   comment = TextAreaField(
     name='comment',
     label=lang.get("table_column_comment").capitalize(),
-    validators=[DataRequired()]
+    validators=[DataRequired()],
+    render_kw=dict(placeholder=lang.get("placeholder_input_comment").capitalize())
   )
   internal = BooleanField(
     name='internal',