+++ ProdManager/routes/service/forms.py
@@ -14,7 +14,8 @@ class ServiceCreateForm(CustomForm):
   description = TextAreaField(
     name='description',
     label=lang.get("table_column_description").capitalize(),
-    validators=[Optional()]
+    validators=[Optional()],
+    render_kw=dict(placeholder=lang.get("placeholder_input_description").capitalize())
   )
 
 class ServiceUpdateForm(ServiceCreateForm):