+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 
 - feat(markdown): Add support for Markdown syntaxe in resources description (!130)
+- feat(forms.\*): Add placeholder in forms textarea inputs (!131)
 
 ## Fixed